+++ Source/Classes/ProjectValidateAttribute.ps1
@@ -4,16 +4,18 @@ class ProjectValidateAttribute : ValidateArgumentsAttribute {
    [void] Validate(
       [object] $arguments,
       [EngineIntrinsics] $EngineIntrinsics) {
-     
-      if (_hasProjectCacheExpired) {
-         [VSTeamProjectCache]::projects = _getProjects
-         [VSTeamProjectCache]::timestamp = (Get-Date).Minute
+
+      # Do not fail on null or empty, leave that to other validation conditions
+      if ([string]::IsNullOrEmpty($arguments)) { 
+         return 
       }
 
-      if (($null -ne [VSTeamProjectCache]::projects) -and (-not ($arguments -in [VSTeamProjectCache]::projects))) {
+      $cachedProjects = [VSTeamProjectCache]::GetCurrent($false)
+
+      if (($cachedProjects.count -gt 0) -and ($arguments -notin $cachedProjects)) {
          throw [ValidationMetadataException]::new(
             "'$arguments' is not a valid project. Valid projects are: '" +
-            ([VSTeamProjectCache]::projects -join "', '") + "'")
+            ($cachedProjects -join "', '") + "'")
       }
    }
 }