+++ Source/Classes/VSTeamVersions.ps1
@@ -1,5 +1,6 @@
 class VSTeamVersions {
    static [string] $Account = $env:TEAM_ACCT
+   static [string] $DefaultTimeout = $env:TEAM_TIMEOUT
    static [string] $DefaultProject = $env:TEAM_PROJECT
    static [string] $DefaultProcess = $env:TEAM_PROCESS
    static [string] $Version = $(If ($env:TEAM_VERSION) { $env:TEAM_VERSION } Else { "TFS2017" })
@@ -18,5 +19,5 @@ class VSTeamVersions {
    static [string] $ModuleVersion = $null
    static [string] $Graph = ''
    static [string] $Policy = '3.0'
-   static [string] $ProcessDefinition = ''
+   static [string] $Processes = ''
 }