+++ Source/Public/Get-VSTeamAPIVersion.ps1
@@ -3,7 +3,7 @@ function Get-VSTeamAPIVersion {
    [OutputType([System.Collections.Hashtable])]
    param(
       [Parameter(Mandatory = $false, Position = 0)]
-      [ValidateSet('Build', 'Release', 'Core', 'Git', 'DistributedTask', 'VariableGroups', 'Tfvc', 'Packaging', 'MemberEntitlementManagement', 'ExtensionsManagement', 'ServiceEndpoints', 'Graph', 'TaskGroups', 'Policy')]
+      [ValidateSet('Build', 'Release', 'Core', 'Git', 'DistributedTask', 'VariableGroups', 'Tfvc', 'Packaging', 'MemberEntitlementManagement', 'ExtensionsManagement', 'ServiceEndpoints', 'Graph', 'TaskGroups', 'Policy', 'Processes')]
       [string] $Service
    )
 
@@ -24,9 +24,10 @@ function Get-VSTeamAPIVersion {
          TaskGroups                  = $(_getApiVersion TaskGroups)
          MemberEntitlementManagement = $(_getApiVersion MemberEntitlementManagement)
          ExtensionsManagement        = $(_getApiVersion ExtensionsManagement)
-         ServiceEndpoints       = $(_getApiVersion ServiceEndpoints)
+         ServiceEndpoints            = $(_getApiVersion ServiceEndpoints)
          Graph                       = $(_getApiVersion Graph)
          Policy                      = $(_getApiVersion Policy)
+         Processes                   = $(_getApiVersion Processes)
       }
    }
 }