+++ unit/test/Add-VSTeamProject.Tests.ps1
@@ -62,12 +62,22 @@ Describe 'VSTeamProject' {
 
       Mock _getInstance { return 'https://dev.azure.com/test' }
       Mock _getApiVersion { return '1.0-unitTests' }
-      Mock _callApi -ParameterFilter {$area -eq 'work' -and $resource -eq 'processes'} -MockWith {
-            return @{value = @(
-                        @{name = 'Agile'; Typeid = '00000000-0000-0000-0000-000000000001' },
-                        @{name = 'CMMI' ; Typeid = '00000000-0000-0000-0000-000000000002' },
-                        @{name = 'Scrum'; Typeid = '00000000-0000-0000-0000-000000000003' }
-            )}
+      Mock _callApi -ParameterFilter { $area -eq 'work' -and $resource -eq 'processes' } -MockWith {
+         return [PSCustomObject]@{value = @(
+               [PSCustomObject]@{
+                  name   = 'Agile'
+                  Typeid = '00000000-0000-0000-0000-000000000001' 
+               },
+               [PSCustomObject]@{
+                  name   = 'CMMI'
+                  Typeid = '00000000-0000-0000-0000-000000000002' 
+               },
+               [PSCustomObject]@{
+                  name   = 'Scrum'
+                  Typeid = '00000000-0000-0000-0000-000000000003' 
+               }
+            )
+         }
       }         
    }
 
@@ -100,7 +110,8 @@ Describe 'VSTeamProject' {
          }
 
       }
-#-Area 'work' -resource 'processes'       
+
+      # -Area 'work' -resource 'processes'       
       It 'with tfvc should create project with tfvc' {
          Add-VSTeamProject -Name Test -tfvc
 
@@ -135,7 +146,12 @@ Describe 'VSTeamProject' {
          Mock Invoke-RestMethod { return @{status = 'inProgress'; id = 1; url = 'https://someplace.com' } } -ParameterFilter { $Method -eq 'Post' -and $Uri -eq "https://dev.azure.com/test/_apis/projects?api-version=$(_getApiVersion Core)" }
          Mock _trackProjectProgress
          Mock Invoke-RestMethod { return $singleResult } -ParameterFilter { $Uri -eq "https://dev.azure.com/test/_apis/projects/Test?api-version=$(_getApiVersion Core)" }
-         Mock Get-VSTeamProcess { return @{name = 'CMMI'; id = 1 } }
+         Mock Get-VSTeamProcess { return [PSCustomObject]@{
+               name   = 'CMMI'
+               id     = 1
+               Typeid = '00000000-0000-0000-0000-000000000002' 
+            } 
+         }
       }
 
       It 'Should create project with CMMI' {