+++ unit/test/ProcessValidateAttribute.Tests.ps1
@@ -11,17 +11,17 @@ Describe "ProcessValidateAttribute" {
       . "$PSScriptRoot/../../Source/Private/common.ps1"
       . "$PSScriptRoot/../../Source/Classes/$sut"
       . "$PSScriptRoot/../../Source/Public/Get-VSTeamProcess.ps1"
-
    }
 
    Context "Existing Process" {
       BeforeAll {
          Mock Get-VSTeamProcess {
-             return @(
-               [PSCustomObject]@{Name = "Test1";            url='http://bogus.none/100'},
-               [PSCustomObject]@{Name = "Test2";            url='http://bogus.none/101'}
-            ) 
+            return @(
+               [PSCustomObject]@{ Name = "Test1"; url = 'http://bogus.none/100' },
+               [PSCustomObject]@{ Name = "Test2"; url = 'http://bogus.none/101' }
+            )
          }
+
          [VSTeamProcessCache]::Invalidate()
       }
 
@@ -32,20 +32,31 @@ Describe "ProcessValidateAttribute" {
          [VSTeamProcessCache]::processes.Count | should -BeGreaterThan 0
          [VSTeamProcessCache]::timestamp       | should -BeGreaterOrEqual 0
       }
+
+      It "should throw if name is not in list " {
+         $target = [ProcessValidateAttribute]::new()
+         { $target.Validate("Test1", $null) } | Should -Not -Throw
+
+         Should -Invoke -CommandName Get-VSTeamProcess -Times 1 -Exactly
+
+         [VSTeamProcessCache]::processes.Count | should -BeGreaterThan 0
+
+         [VSTeamProcessCache]::timestamp | should -BeGreaterOrEqual 0
+      }
+
       It "should throw if name is not in list " {
          $target = [ProcessValidateAttribute]::new()
 
          { $target.Validate("Test", $null) } | Should -Throw
       }
-
-
    }
 
    Context "No Processes" {
       BeforeAll {
          Mock Get-VSTeamProcess {
-             return @(  ) 
+            return @()
          }
+
          [VSTeamProcessCache]::Invalidate()
       }